requirements.txt
attrs==21.2.0
Automat==20.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.7
cssselect==1.1.0
cycler==0.10.0
Flask==2.0.1
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
itemadapter==0.3.0
itemloaders==1.0.4
itsdangerous==2.0.1
jieba==0.42.1
Jinja2==3.0.1
jmespath==0.10.0
kiwisolver==1.3.1
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib==3.4.2
numpy==1.21.1
parsel==1.6.0
Pillow==8.3.1
priority==1.3.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.2
queuelib==1.6.1
requests==2.26.0
Scrapy==2.5.0
service-identity==21.1.0
six==1.16.0
soupsieve==2.2.1
Twisted==21.2.0
twisted-iocpsupport==1.0.1
urllib3==1.26.6
w3lib==1.22.0
Werkzeug==2.0.1
wordcloud @ file:///D:/wordcloud-1.8.1-cp39-cp39-win_amd64.whl
xlwt==1.3.0
zope.interface==5.4.0
